+++ Sources/WalletConnectRelay/ClientAuth/ClientIdStorage.swift
@@ -39,76 +39,3 @@ public struct ClientIdStorage: ClientIdStoring {
         clientIdMigrationController?.migrateIfNeeded()
     }
 }
-
-public class ClientIdMigrationController {
-
-    private let service: String
-
-    private let secItem: KeychainServiceProtocol
-
-    private let keyValueStorage: KeyValueStorage
-
-    private let lastMigrationKey = "com.walletconnect.iridium.last_client_id_migration"
-
-    private let cleintIdStorageKey = "com.walletconnect.iridium.client_id"
-
-    public init(
-        keychainService: KeychainServiceProtocol = KeychainServiceWrapper(),
-        serviceIdentifier: String,
-        keyValueStorage: KeyValueStorage,
-        logger: ConsoleLogging
-    ) {
-        self.secItem = keychainService
-        self.keyValueStorage = keyValueStorage
-        service = serviceIdentifier
-    }
-
-    func migrateIfNeeded() {
-        getClientId()
-        if let lastMigration = keyValueStorage.object(forKey: lastMigrationKey) as? String {
-            print("last migration: \(lastMigration)")
-            return
-        }
-
-        print("ClientIdMigrationController: Migrating client id")
-        let query: [CFString: Any] = [
-            kSecClass: kSecClassGenericPassword,
-            kSecAttrAccessible: kSecAttrAccessibleWhenUnlockedThisDeviceOnly,
-            kSecAttrIsInvisible: true,
-            kSecUseDataProtectionKeychain: true,
-            kSecAttrService: service,
-            kSecAttrAccount: cleintIdStorageKey
-        ]
-        let attributes = [kSecAttrAccessible: kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly]
-
-        let status = secItem.update(query as CFDictionary, attributes as CFDictionary)
-        print(status.message)
-        guard status == errSecSuccess else {
-            print(status)
-            return
-        }
-        getClientId()
-
-
-        keyValueStorage.set(EnvironmentInfo.packageVersion, forKey: lastMigrationKey)
-    }
-
-    private func getClientId() {
-        var getQuery: [CFString: Any] = [
-            kSecClass: kSecClassGenericPassword,
-            kSecAttrAccessible: kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly,
-            kSecAttrIsInvisible: true,
-            kSecUseDataProtectionKeychain: true,
-            kSecAttrService: service,
-            kSecAttrAccount: cleintIdStorageKey
-        ]
-
-        getQuery[kSecReturnData] = true
-
-        var item: CFTypeRef?
-        let getStatus = secItem.copyMatching(getQuery as CFDictionary, &item)
-        print(getStatus.description)
-
-    }
-
-}
